StashItem class renamed to CachedItem. It has it's own file now.
